pulled out sharding logic into db config
