+++ v2/sourcedb-to-spanner/src/main/java/com/google/cloud/teleport/v2/templates/SourceDbToSpanner.java
@@ -109,7 +109,7 @@ public class SourceDbToSpanner {
       return PipelineController.executeShardedMigration(options, pipeline, shards, spannerConfig);
     } else {
       // return PipelineController.executeSingleInstanceMigration(options, pipeline, spannerConfig);
-      return PipelineController.executeMigration(options, pipeline, spannerConfig);
+      return PipelineController.executeSingleInstanceMigration(options, pipeline, spannerConfig);
     }
   }
